+++ Recipe.h
@@ -4,12 +4,12 @@ class Recipe
 {
 private:
     unsigned int _stages;
-    unsigned int *_pauses;
+    unsigned long *_pauses;
     unsigned int *_temperatures;
 public:
     Recipe(unsigned int stage_count);
     ~Recipe();
-    void setStage(unsigned int stage, unsigned int pause, unsigned int temperature);
+    void setStage(unsigned int stage, unsigned long pause, unsigned int temperature);
     unsigned int getTemperature(unsigned int stage);
     unsigned int getPause(unsigned int stage);
     unsigned int getStageCount();
